refactor all repos and optimize some libs to reduce app-bitcoin size to 750
