+++ .vscode/settings.json
@@ -1,8 +1,9 @@
 {
   "rust-analyzer.linkedProjects": [
-    "./contracts/cw-bitcoin/Cargo.toml",
+    "./contracts/app-bitcoin/Cargo.toml",
     "./contracts/light-client-bitcoin/Cargo.toml",
     "./packages/common-bitcoin/Cargo.toml",
+    "./packages/light-client/Cargo.toml",
     "./Cargo.toml"
   ],
   "terminal.integrated.defaultProfile.linux": "zsh"
