+++ contracts/app-bitcoin/Cargo.toml
@@ -4,7 +4,7 @@ documentation = {workspace = true}
 edition = {workspace = true}
 homepage = {workspace = true}
 license = {workspace = true}
-name = "cw-light-client-bitcoin"
+name = "cw-app-bitcoin"
 repository = {workspace = true}
 version = {workspace = true}
 
@@ -14,27 +14,39 @@ crate-type = ["cdylib", "rlib"]
 
 [features]
 # use library feature to disable all instantiate/execute/query exports
-default = []
+default = ['mainnet']
 library = []
+mainnet = []
 test-tube = []
 
 [dependencies]
+arrayref = "0.3"
 bitcoin = {workspace = true, features = ["serde"]}
-chrono = {workspace = true}
+bitcoin-script = {workspace = true}
 common-bitcoin = {workspace = true}
 cosmwasm-schema = {workspace = true}
 cosmwasm-std = {workspace = true}
 cw-storage-plus = {workspace = true}
 cw2 = {workspace = true}
+cw20 = {workspace = true}
+cw20-base = {workspace = true}
 derive_more = {workspace = true}
+libsecp256k1-core = "0.3.0"
 light-client-bitcoin = {workspace = true}
+oraiswap = {workspace = true}
+prost = {workspace = true}
+sha2 = {workspace = true}
+skip = {workspace = true}
 thiserror = {workspace = true}
+token-bindings = {workspace = true}
+tokenfactory = {workspace = true, features = ["library"]}
 
 [dev-dependencies]
 bitcoincore-rpc-async = {workspace = true}
 bitcoind = {workspace = true}
 chrono = {workspace = true}
 cosmwasm-testing-util = {workspace = true}
+cw-light-client-bitcoin = {workspace = true}
 serde = {workspace = true}
 serde_json = {workspace = true}
 serial_test = "3.1.1"
