+++ contracts/app-bitcoin/src/app.rs
