+++ contracts/app-bitcoin/src/checkpoint.json
