+++ contracts/app-bitcoin/src/checkpoint.rs
