+++ contracts/app-bitcoin/src/constants.rs
