+++ contracts/app-bitcoin/src/contract.rs
