+++ contracts/app-bitcoin/src/entrypoints/execute.rs
