+++ contracts/app-bitcoin/src/entrypoints/mod.rs
