+++ contracts/app-bitcoin/src/entrypoints/query.rs
