+++ contracts/app-bitcoin/src/helper.rs
