+++ contracts/app-bitcoin/src/integration_tests/utils.rs
