+++ contracts/app-bitcoin/src/interface.rs
