+++ contracts/app-bitcoin/src/outpoint_set.rs
