+++ contracts/app-bitcoin/src/recovery.rs
