+++ contracts/app-bitcoin/src/signatory.rs
