+++ contracts/app-bitcoin/src/state.rs
