+++ contracts/app-bitcoin/src/tests/signatory.rs
