+++ contracts/app-bitcoin/src/tests/testdata/cw-bitcoin.wasm
