+++ contracts/app-bitcoin/src/tests/testdata/light-client-bitcoin.wasm
