+++ contracts/app-bitcoin/src/threshold_sig.rs
