+++ contracts/light-client-bitcoin/src/lib.rs
@@ -1,12 +1,10 @@
 pub mod contract;
 pub mod header;
-pub mod msg;
 
 mod constants;
 mod entrypoints;
 #[cfg(test)]
 mod integration_tests;
-mod interface;
 mod state;
 #[cfg(test)]
 mod tests;
