+++ packages/common-bitcoin/Cargo.toml
@@ -26,7 +26,6 @@ cosmwasm-schema = {workspace = true}
 cosmwasm-std = {workspace = true}
 cw-storage-plus = {workspace = true}
 derive_more = {workspace = true}
-ibc-proto = {workspace = true}
 libsecp256k1-core = "0.3.0"
 thiserror = {workspace = true}
