+++ packages/light-client/Cargo.toml
@@ -1,10 +1,12 @@
 [package]
 authors = {workspace = true}
+description = "Common light client bitcoin types"
 documentation = {workspace = true}
 edition = {workspace = true}
 homepage = {workspace = true}
 license = {workspace = true}
-name = "cw-bitcoin"
+name = "light-client-bitcoin"
+readme = "README.md"
 repository = {workspace = true}
 version = {workspace = true}
 
@@ -14,10 +16,7 @@ crate-type = ["cdylib", "rlib"]
 
 [features]
 # use library feature to disable all instantiate/execute/query exports
-default = ['mainnet']
-library = []
-mainnet = []
-test-tube = []
+default = []
 
 [dependencies]
 arrayref = "0.3"
@@ -26,27 +25,15 @@ bitcoin-script = {workspace = true}
 common-bitcoin = {workspace = true}
 cosmwasm-schema = {workspace = true}
 cosmwasm-std = {workspace = true}
-cw-storage-plus = {workspace = true}
-cw2 = {workspace = true}
-cw20 = {workspace = true}
-cw20-base = {workspace = true}
 derive_more = {workspace = true}
 libsecp256k1-core = "0.3.0"
-light-client-bitcoin = {workspace = true, features = ["library"]}
-oraiswap = {workspace = true}
-prost = {workspace = true}
-sha2 = {workspace = true}
-skip = {workspace = true}
 thiserror = {workspace = true}
-token-bindings = {workspace = true}
-tokenfactory = {workspace = true, features = ["library"]}
 
 [dev-dependencies]
 bitcoincore-rpc-async = {workspace = true}
 bitcoind = {workspace = true}
-chrono = {workspace = true}
-cosmwasm-testing-util = {workspace = true}
 serde = {workspace = true}
 serde_json = {workspace = true}
+
+chrono = "0.4.19"
 serial_test = "3.1.1"
-tokio = {workspace = true}
