+++ packages/light-client/src/checkpoint.json
